azure-speech-service-text-to-speech-using-ssml/requirements.txt
# Azure Text-to-Speech Dependencies
azure-cognitiveservices-speech==1.34.0
python-dotenv==1.0.0
asyncio
